requirements.txt
asgiref==3.6.0
certifi==2022.12.7
deprecation==2.1.0
Django==3.2.18
django-email-verification==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
packaging==23.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
sentry-sdk==1.19.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.15
